[minimega] Recover VMs still running after a minimega crash

+++ cmd/minimega/main.go
@@ -41,6 +41,7 @@ var (
 	f_broadcastIP = flag.String("broadcast", "255.255.255.255", "meshage broadcast address to use")
 	f_port        = flag.Int("port", 9000, "meshage port to listen on")
 	f_force       = flag.Bool("force", false, "force minimega to run even if it appears to already be running")
+	f_recover     = flag.Bool("recover", false, "attempt to recover from a previously running instance (only if -force is not set)")
 	f_nostdin     = flag.Bool("nostdin", false, "disable reading from stdin, useful for putting minimega in the background")
 	f_version     = flag.Bool("version", false, "print the version and copyright notices")
 	f_context     = flag.String("context", "minimega", "meshage context for discovery")
@@ -196,13 +197,20 @@ func main() {
 
 	// check for a running instance of minimega
 	if _, err := os.Stat(filepath.Join(*f_base, "minimega")); err == nil {
-		if !*f_force {
-			log.Fatalln("minimega appears to already be running, override with -force")
-		}
-		log.Warn("minimega may already be running, proceed with caution")
+		if *f_force {
+			log.Warn("minimega may already be running, proceed with caution")
 
-		if err := os.Remove(filepath.Join(*f_base, "minimega")); err != nil {
-			log.Fatalln(err)
+			if err := os.Remove(filepath.Join(*f_base, "minimega")); err != nil {
+				log.Fatalln(err)
+			}
+		} else if *f_recover {
+			if err := os.Remove(filepath.Join(*f_base, "minimega")); err != nil {
+				log.Fatalln(err)
+			}
+
+			// an attempt to recover will happen later after the mesh is initialized
+		} else {
+			log.Fatalln("minimega appears to already be running, override with -force")
 		}
 	}
 
@@ -237,6 +245,12 @@ func main() {
 
 	if err := setupMeshageLogging(*f_headnode); err != nil {
 		log.Fatal("unable to setup mesh logging: %v", err)
+	}
+
+	if *f_recover { // has to happen after meshageNode is created
+		if err := recover(); err != nil {
+			log.Fatal("recovery failed: %v", err)
+		}
 	}
 
 	// has to happen after meshageNode is created
